fast-swiftches, function reorganization
git-svn-id: http://svn.seqan.de/seqan/trunk@6934 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/index/find_swift.h
@@ -872,7 +872,7 @@ inline bool _swiftMultiProcessQGram(
 		TShortSize hitCount;
 
 		do {
-			if ((__int64)(*bkt).lastIncrement < bktBeginHstk + pattern.finderPosOffset
+			if ((__int64)(*bkt).lastIncrement < bktBeginHstk + (__int64)pattern.finderPosOffset
 				|| (__int64)((*bkt).lastIncrement + bucketParams.distanceCut) < (__int64)(curPos + length(pattern.shape)))
 			{
 				// last increment was before the beginning of the current bucket => bucket is reused
